--- lora_tune/results/results_0_5.json
@@ -1,188 +1,0 @@
-[
-  {
-    "name": "Qwen2.5-0.5B Base (codelm_leftright_context)",
-    "languages": {
-      "python": {
-        "EM": 0.0,
-        "ES": 18.69,
-        "ES RepoEval": 11.28
-      },
-      "java": {
-        "EM": 0.0,
-        "ES": 4.64,
-        "ES RepoEval": 2.78
-      },
-      "csharp": {
-        "EM": 0.0,
-        "ES": 6.93,
-        "ES RepoEval": 4.21
-      },
-      "typescript": {
-        "EM": 0.06,
-        "ES": 16.64,
-        "ES RepoEval": 10.0
-      }
-    },
-    "overall": {
-      "EM": 0.02,
-      "ES": 12.88,
-      "ES RepoEval": 7.76
-    },
-    "source_file": "0_5_results.md"
-  },
-  {
-    "name": "Qwen2.5-0.5B Base EPOCH 1 (QLora r=32, a=64, code_x_glue (Python+Java)) (codelm_leftright_context)",
-    "languages": {
-      "python": {
-        "EM": 0.15,
-        "ES": 25.72,
-        "ES RepoEval": 15.9
-      },
-      "java": {
-        "EM": 1.59,
-        "ES": 34.59,
-        "ES RepoEval": 24.06
-      },
-      "csharp": {
-        "EM": 0.62,
-        "ES": 35.49,
-        "ES RepoEval": 24.02
-      },
-      "typescript": {
-        "EM": 0.51,
-        "ES": 31.96,
-        "ES RepoEval": 20.99
-      }
-    },
-    "overall": {
-      "EM": 0.67,
-      "ES": 31.48,
-      "ES RepoEval": 20.82
-    },
-    "source_file": "0_5_results.md"
-  },
-  {
-    "name": "Qwen2.5-0.5B Base EPOCH 2 (QLora r=32, a=64, code_x_glue (Python+Java)) (codelm_leftright_context)",
-    "languages": {
-      "python": {
-        "EM": 0.08,
-        "ES": 23.39,
-        "ES RepoEval": 14.22
-      },
-      "java": {
-        "EM": 0.37,
-        "ES": 30.57,
-        "ES RepoEval": 20.05
-      },
-      "csharp": {
-        "EM": 0.11,
-        "ES": 32.48,
-        "ES RepoEval": 21.3
-      },
-      "typescript": {
-        "EM": 0.09,
-        "ES": 31.25,
-        "ES RepoEval": 20.12
-      }
-    },
-    "overall": {
-      "EM": 0.15,
-      "ES": 29.21,
-      "ES RepoEval": 18.73
-    },
-    "source_file": "0_5_results.md"
-  },
-  {
-    "name": "Qwen2.5-0.5B Base EPOCH 3 (QLora r=32, a=64, code_x_glue (Python+Java)) (codelm_leftright_context)",
-    "languages": {
-      "python": {
-        "EM": 0.04,
-        "ES": 22.73,
-        "ES RepoEval": 13.71
-      },
-      "java": {
-        "EM": 0.33,
-        "ES": 29.85,
-        "ES RepoEval": 19.31
-      },
-      "csharp": {
-        "EM": 0.06,
-        "ES": 31.28,
-        "ES RepoEval": 20.35
-      },
-      "typescript": {
-        "EM": 0.12,
-        "ES": 30.28,
-        "ES RepoEval": 19.54
-      }
-    },
-    "overall": {
-      "EM": 0.13,
-      "ES": 28.34,
-      "ES RepoEval": 18.07
-    },
-    "source_file": "0_5_results.md"
-  },
-  {
-    "name": "Qwen2.5-0.5B Base EPOCH 4 (QLora r=32, a=64, code_x_glue (Python+Java)) (codelm_leftright_context)",
-    "languages": {
-      "python": {
-        "EM": 0.04,
-        "ES": 22.71,
-        "ES RepoEval": 13.73
-      },
-      "java": {
-        "EM": 0.33,
-        "ES": 29.75,
-        "ES RepoEval": 19.32
-      },
-      "csharp": {
-        "EM": 0.06,
-        "ES": 31.07,
-        "ES RepoEval": 20.39
-      },
-      "typescript": {
-        "EM": 0.12,
-        "ES": 30.58,
-        "ES RepoEval": 19.62
-      }
-    },
-    "overall": {
-      "EM": 0.13,
-      "ES": 28.45,
-      "ES RepoEval": 18.12
-    },
-    "source_file": "0_5_results.md"
-  },
-  {
-    "name": "Qwen2.5-0.5B Base EPOCH 5 (QLora r=32, a=64, code_x_glue (Python+Java)) (codelm_leftright_context)",
-    "languages": {
-      "python": {
-        "EM": 0.04,
-        "ES": 22.76,
-        "ES RepoEval": 13.76
-      },
-      "java": {
-        "EM": 0.33,
-        "ES": 29.99,
-        "ES RepoEval": 19.36
-      },
-      "csharp": {
-        "EM": 0.06,
-        "ES": 31.4,
-        "ES RepoEval": 20.42
-      },
-      "typescript": {
-        "EM": 0.12,
-        "ES": 30.97,
-        "ES RepoEval": 19.97
-      }
-    },
-    "overall": {
-      "EM": 0.13,
-      "ES": 28.63,
-      "ES RepoEval": 18.25
-    },
-    "source_file": "0_5_results.md"
-  }
-]
