+++ tuning/custom_recipes/fim/fim_config.example.yaml
