+++ tuning/lora_tune/config.yaml
@@ -62,14 +62,19 @@ resume_from_checkpoint: False
 
 dataset:
   # Use the custom builder function from the modified file
-  _component_: custom_datasets.text_completion_dataset
+  _component_: tuning.custom_datasets.codexglue.text_completion_dataset
   # source, column, split, add_eos etc are arguments to the builder function
   source: google/code_x_glue_cc_code_completion_token
   column: code
-  add_eos: True
   split: train[:100%]
   packed: True # VERY IMPORTANT
-  fim_prob: 1.0
+
+make_fim:
+  convert_to_fim: True # whether to convert dataset to FIM
+  fim_prob: 1.0 # prob a single sample is converted to FIM
+  min_fim_middle_percent: 0.01 # min length of middle portion as percent of total length
+  max_fim_middle_percent: 0.1 # max length of middle portion as percent of total length
+  verbose: False # whether to include debug output about the FIM-transformation process
 
 seed: null
 shuffle: True
