+++ tuning/lora_tune/run_lora.sh
@@ -1,0 +1,7 @@
+#!/bin/bash
+
+export PYTHONPATH=$PYTHONPATH:../..
+
+tune download Qwen/Qwen2.5-0.5B --output-dir /tmp/Qwen2_5-0_5B-Base
+
+tune run tuning/custom_recipes/fim/fim_lora_recipe.py --config config.yaml
